Refactor OpenTelemetry configuration
This commit moves the OpenTelemetry initialization from the frontend
module to the internal module to make it reusable by the backend module.
