+++ internal/tracing/otel_sdk.go
@@ -4,17 +4,15 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+package tracing
 
-package frontend
-
-// TODO: can we move this into ocm-sdk-go/tracing?
 import (
 	"context"
 	"fmt"
@@ -26,8 +24,6 @@ import (
 	"go.opentelemetry.io/otel/propagation"
 	"go.opentelemetry.io/otel/sdk/trace"
 	semconv "go.opentelemetry.io/otel/semconv/v1.27.0"
-
-	"github.com/Azure/ARO-HCP/internal/version"
 
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/sdk/resource"
@@ -69,11 +65,8 @@ func ConfigureOpenTelemetryTracer(ctx context.Context, logger *slog.Logger, reso
 	if len(resourceAttrs) > 0 {
 		opts = append(opts, resource.WithAttributes(resourceAttrs...))
 	}
-	opts = append(opts, resource.WithAttributes(
-		semconv.ServiceNameKey.String(ProgramName),
-		semconv.ServiceVersionKey.String(version.CommitSHA),
-		semconv.CloudProviderAzure,
-	))
+	opts = append(opts, resource.WithAttributes(semconv.CloudProviderAzure))
+
 	resources, err := resource.New(ctx, opts...)
 	if err != nil {
 		return nil, fmt.Errorf("failed to initialise trace resources: %w", err)
